+++ README.md
@@ -45,7 +45,7 @@
 
 # `STEP5 - 자동차 경주(리팩토링)`
 - [X] ...... READEME 작성
-- [ ] ...... 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현
+- [X] ...... 핵심 비지니스 로직을 가지는 객체를 domain 패키지, UI 관련한 객체를 view 패키지에 구현
 - [ ] ...... MVC 패턴 기반으로 리팩토링해 view 패키지의 객체가 domain 패키지 객체에 의존할 수 있지만, domain 패키지의 객체는 view 패키지 객체에 의존하지 않도록 구현
 
 # `리뷰 반영 사항`
