+++ src/main/java/racing/domain/Count.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 public class Count {
