+++ src/main/java/racing/domain/RacingHistory.java
@@ -1,4 +1,4 @@
-package racing;
+package racing.domain;
 
 import java.util.Set;
