+++ src/main/java/racing/view/InputView.java
@@ -1,4 +1,4 @@
-package ui;
+package racing.view;
 
 import java.util.Scanner;
