+++ src/test/java/racing/RacingServiceTest.java
@@ -5,6 +5,8 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+import racing.domain.Car;
+import racing.domain.Location;
 import util.RacingValidator;
 import util.RandomNumberGenerator;
